chore: fix type error and template change
