+++ packages/create-valaxy/template/valaxy.config.ts
@@ -1,155 +1,13 @@
-import { defineSite } from 'valaxy'
-import type { ThemeConfig } from 'valaxy-theme-yun'
+import { defineConfig } from 'valaxy'
+
+// add icons what you will need
+const safelist = [
+  'i-ri-home-line',
+]
 
 /**
  * User Config
  */
-export default defineSite<ThemeConfig>({
-  lang: 'zh-CN',
-  title: 'Valaxy Theme Yun',
-  author: {
-    name: '云游君',
-  },
-  description: 'Valaxy Theme Yun Preview.',
-  social: [
-    {
-      name: 'RSS',
-      link: '/atom.xml',
-      icon: 'i-ri-rss-line',
-      color: 'orange',
-    },
-    {
-      name: 'QQ 群 1050458482',
-      link: 'https://qm.qq.com/cgi-bin/qm/qr?k=kZJzggTTCf4SpvEQ8lXWoi5ZjhAx0ILZ&jump_from=webapi',
-      icon: 'i-ri-qq-line',
-      color: '#12B7F5',
-    },
-    {
-      name: 'GitHub',
-      link: 'https://github.com/YunYouJun',
-      icon: 'i-ri-github-line',
-      color: '#6e5494',
-    },
-    {
-      name: '微博',
-      link: 'https://weibo.com/jizhideyunyoujun',
-      icon: 'i-ri-weibo-line',
-      color: '#E6162D',
-    },
-    {
-      name: '豆瓣',
-      link: 'https://www.douban.com/people/yunyoujun/',
-      icon: 'i-ri-douban-line',
-      color: '#007722',
-    },
-    {
-      name: '网易云音乐',
-      link: 'https://music.163.com/#/user/home?id=247102977',
-      icon: 'i-ri-netease-cloud-music-line',
-      color: '#C20C0C',
-    },
-    {
-      name: '知乎',
-      link: 'https://www.zhihu.com/people/yunyoujun/',
-      icon: 'i-ri-zhihu-line',
-      color: '#0084FF',
-    },
-    {
-      name: '哔哩哔哩',
-      link: 'https://space.bilibili.com/1579790',
-      icon: 'i-ri-bilibili-line',
-      color: '#FF8EB3',
-    },
-    {
-      name: '微信公众号',
-      link: 'https://cdn.yunyoujun.cn/img/about/white-qrcode-and-search.jpg',
-      icon: 'i-ri-wechat-2-line',
-      color: '#1AAD19',
-    },
-    {
-      name: 'Twitter',
-      link: 'https://twitter.com/YunYouJun',
-      icon: 'i-ri-twitter-line',
-      color: '#1da1f2',
-    },
-    {
-      name: 'Telegram Channel',
-      link: 'https://t.me/elpsycn',
-      icon: 'i-ri-telegram-line',
-      color: '#0088CC',
-    },
-    {
-      name: 'E-Mail',
-      link: 'mailto:me@yunyoujun.cn',
-      icon: 'i-ri-mail-line',
-      color: '#8E71C1',
-    },
-    {
-      name: 'Travelling',
-      link: 'https://travellings.link',
-      icon: 'i-ri-train-line',
-      color: 'var(--va-c-text)',
-    },
-  ],
-
-  search: {
-    enable: false,
-  },
-
-  sponsor: {
-    enable: true,
-    title: '我很可爱，请给我钱！',
-    methods: [
-      {
-        name: '支付宝',
-        url: 'https://cdn.yunyoujun.cn/img/donate/alipay-qrcode.jpg',
-        color: '#00A3EE',
-        icon: 'i-ri-alipay-line',
-      },
-      {
-        name: 'QQ 支付',
-        url: 'https://cdn.yunyoujun.cn/img/donate/qqpay-qrcode.png',
-        color: '#12B7F5',
-        icon: 'i-ri-qq-line',
-      },
-      {
-        name: '微信支付',
-        url: 'https://cdn.yunyoujun.cn/img/donate/wechatpay-qrcode.jpg',
-        color: '#2DC100',
-        icon: 'i-ri-wechat-pay-line',
-      },
-    ],
-  },
-
-  theme: 'yun',
-
-  themeConfig: {
-    banner: {
-      enable: true,
-      title: '云游君的小站',
-    },
-
-    pages: [
-      {
-        name: '我的小伙伴们',
-        url: '/links/',
-        icon: 'i-ri-genderless-line',
-        color: 'dodgerblue',
-      },
-      {
-        name: '喜欢的女孩子',
-        url: '/girls/',
-        icon: 'i-ri-women-line',
-        color: 'hotpink',
-      },
-    ],
-
-    footer: {
-      since: 2016,
-      beian: {
-        enable: true,
-        icp: '苏ICP备17038157号',
-      },
-    },
-  },
+export default defineConfig({
+  unocss: { safelist },
 })
