+++ packages/valaxy/client/modules/valaxy.ts
@@ -12,7 +12,7 @@ import { useStorage } from '@vueuse/core'
 
 import type { Router } from 'vue-router'
 import type { PageDataPayload } from '../../types'
-import { initSite, ValaxySiteConfigSymbol } from '../config'
+import { ValaxySiteConfigSymbol, initSite } from '../config'
 import { ensureSuffix } from '@antfu/utils'
 
 import type { UserModule } from 'valaxy/client/types'
